adjust gap logic
